requirements.txt
pandas
flask
flasgger
langchain
langchain-google-genai
python-dotenv
 nltk
 rouge-score
 